Upgrade KC version to 17 and adapted themes
