+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# CHANGELOG
+
+- Upgrade Keycloak to version 21
+- Added CI builds
+- Upgrade to Java 17
+- Used native client for api calls
+- Extended base login
