+++ README.md
@@ -1,1 +1,5 @@
 # Login By PhoneNumber
+
+This plugin helps you login using phon number tan. For that, we create and manage a user attribute `phoneNumber`.
+
+After you install it, you can configure each step independently
