+++ docker-compose.yml
@@ -1,0 +1,19 @@
+version: "3.9"
+
+services:
+  keycloak:
+    build:
+      context: .
+      dockerfile: Dev.Dockerfile
+    ports:
+      - '9100:9100'
+    environment:
+      KEYCLOAK_ADMIN: admin
+      KEYCLOAK_ADMIN_PASSWORD: password
+      KEYCLOAK_FRONTEND_URL: http://localhost:9100/auth
+      KC_HTTP_PORT: 9100
+      KC_METRICS_ENABLED: 'true'
+      KC_LOG_CONSOLE_COLOR: 'true'
+      KC_HEALTH_ENABLED: 'true'
+    command:
+      - start-dev
