--- src/main/java/com/bayamsell/keycloak/authenticator/PhoneNumberChooseUser.java
@@ -1,190 +1,0 @@
-package com.bayamsell.keycloak.authenticator;
-
-import com.bayamsell.keycloak.constants.ConfigKey;
-import com.bayamsell.keycloak.constants.PhoneKey;
-import com.bayamsell.keycloak.constants.Utils;
-import lombok.NoArgsConstructor;
-import lombok.extern.jbosslog.JBossLog;
-import org.keycloak.Config;
-import org.keycloak.authentication.AuthenticationFlowContext;
-import org.keycloak.authentication.Authenticator;
-import org.keycloak.authentication.AuthenticatorFactory;
-import org.keycloak.events.Errors;
-import org.keycloak.events.EventBuilder;
-import org.keycloak.models.AuthenticationExecutionModel;
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
-import org.keycloak.models.RealmModel;
-import org.keycloak.models.UserModel;
-import org.keycloak.models.UserProvider;
-import org.keycloak.provider.ProviderConfigProperty;
-import org.keycloak.provider.ServerInfoAwareProviderFactory;
-import org.keycloak.sessions.AuthenticationSessionModel;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static com.bayamsell.keycloak.constants.PhoneNumber.DEFAULT_PHONE_KEY_NAME;
-import static com.bayamsell.keycloak.constants.PhoneNumber.handleConfiguredFor;
-
-@JBossLog
-@NoArgsConstructor
-public class PhoneNumberChooseUser implements
-        Authenticator,
-        AuthenticatorFactory,
-        ServerInfoAwareProviderFactory {
-
-    public static final String PROVIDER_ID = "phone-number-choose-user";
-    public static final AuthenticationExecutionModel.Requirement[] REQUIREMENT_CHOICES = {
-            AuthenticationExecutionModel.Requirement.REQUIRED
-    };
-
-    private static final List<ProviderConfigProperty> configProperties = new ArrayList<>();
-    private static final Map<String, String> infos = new HashMap<>();
-
-    static {
-        infos.put("version", "1.0.0");
-
-        ProviderConfigProperty property;
-
-        // Phone name
-        property = new ProviderConfigProperty();
-        property.setName(ConfigKey.USER_PHONE_ATTRIBUTE_NAME);
-        property.setLabel("Phone attribute name");
-        property.setType(ProviderConfigProperty.STRING_TYPE);
-        property.setHelpText("User phone names's attribute");
-        configProperties.add(property);
-    }
-
-    @Override
-    public void authenticate(AuthenticationFlowContext context) {
-        EventBuilder event = context.getEvent();
-        AuthenticationSessionModel authenticationSession = context.getAuthenticationSession();
-        String phoneNumber = authenticationSession.getAuthNote(PhoneKey.ATTEMPTED_PHONE_NUMBER);
-
-        RealmModel realm = context.getRealm();
-        String attrName = Utils
-                .getConfigString(context.getAuthenticatorConfig(),
-                        ConfigKey.USER_PHONE_ATTRIBUTE_NAME, DEFAULT_PHONE_KEY_NAME);
-        UserModel user = context.getUser();
-
-        if (user != null && !user.isEnabled()) {
-            event.clone()
-                    .detail("phone_number", phoneNumber)
-                    .user(user)
-                    .error(Errors.USER_DISABLED);
-            context.clearUser();
-            context.resetFlow();
-            return;
-        }
-
-        if (user != null && user.isEnabled()) {
-            user.setAttribute(attrName, Collections.singletonList(phoneNumber));
-        } else {
-            UserProvider userProvider = context.getSession().users();
-            List<UserModel> users = userProvider
-                    .searchForUserByUserAttribute(attrName, phoneNumber, realm);
-
-            if (users.size() == 0 && user == null) {
-                UserModel newUser = userProvider.addUser(realm, phoneNumber);
-                newUser.setAttribute(attrName, Collections.singletonList(phoneNumber));
-                newUser.setEnabled(true);
-                user = newUser;
-            } else {
-                user = users.get(0);
-            }
-        }
-
-        context.setUser(user);
-        context
-                .getAuthenticationSession()
-                .setAuthNote(PhoneKey.ATTEMPTED_PHONE_NUMBER, phoneNumber);
-
-        context.success();
-    }
-
-    @Override
-    public void action(AuthenticationFlowContext context) {
-    }
-
-    @Override
-    public boolean requiresUser() {
-        return false;
-    }
-
-    @Override
-    public boolean configuredFor(KeycloakSession session, RealmModel realm, UserModel user) {
-        return handleConfiguredFor(user);
-    }
-
-    @Override
-    public void setRequiredActions(KeycloakSession session, RealmModel realm, UserModel user) {
-
-    }
-
-    @Override
-    public String getDisplayType() {
-        return "SMS -3 Choose user by Phone number";
-    }
-
-    @Override
-    public String getReferenceCategory() {
-        return "phone-choose";
-    }
-
-    @Override
-    public boolean isConfigurable() {
-        return true;
-    }
-
-    @Override
-    public AuthenticationExecutionModel.Requirement[] getRequirementChoices() {
-        return REQUIREMENT_CHOICES;
-    }
-
-    @Override
-    public boolean isUserSetupAllowed() {
-        return false;
-    }
-
-    @Override
-    public String getHelpText() {
-        return "Choose a user, by his/her phone number, to reset credentials for";
-    }
-
-    @Override
-    public List<ProviderConfigProperty> getConfigProperties() {
-        return configProperties;
-    }
-
-    @Override
-    public Authenticator create(KeycloakSession session) {
-        return this;
-    }
-
-    @Override
-    public void init(Config.Scope config) {
-    }
-
-    @Override
-    public void postInit(KeycloakSessionFactory factory) {
-    }
-
-    @Override
-    public void close() {
-    }
-
-    @Override
-    public String getId() {
-        return PROVIDER_ID;
-    }
-
-    @Override
-    public Map<String, String> getOperationalInfo() {
-        return infos;
-    }
-
-}
