--- src/main/java/com/bayamsell/keycloak/constants/PhoneKey.java
@@ -1,8 +1,0 @@
-package com.bayamsell.keycloak.constants;
-
-public class PhoneKey {
-
-    public static final String ATTEMPTED_PHONE_NUMBER = "ATTEMPTED_PHONE_NUMBER";
-
-    public static final String ATTEMPTED_CODE_NAME = "ATTEMPTED_CODE_NAME";
-}
