+++ src/main/java/com/vymalo/keycloak/constants/ConfigKey.java
@@ -1,0 +1,15 @@
+package com.vymalo.keycloak.constants;
+
+public class ConfigKey {
+    public static final String CONF_PRP_SMS_CODE_LENGTH = "sms-api.attr.length";
+    public static final String CONF_PRP_SMS_URL = "sms-api.attr.url";
+
+    public static final String BASIC_AUTH_USERNAME = "sms-api.auth.basic.username";
+    public static final String BASIC_AUTH_PASSWORD = "sms-api.auth.basic.password";
+
+
+    public static final String USER_PHONE_ATTRIBUTE_NAME = "user.attribute.phone.name";
+
+    private ConfigKey() {
+    }
+}
