+++ src/main/java/com/vymalo/keycloak/constants/PhoneKey.java
@@ -1,0 +1,11 @@
+package com.vymalo.keycloak.constants;
+
+public class PhoneKey {
+
+    public static final String ATTEMPTED_PHONE_NUMBER = "ATTEMPTED_PHONE_NUMBER";
+
+    public static final String ATTEMPTED_CODE_NAME = "ATTEMPTED_CODE_NAME";
+
+    private PhoneKey() {
+    }
+}
