+++ src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,6 +1,6 @@
-com.bayamsell.keycloak.authenticator.PhoneNumberGetNumber
-com.bayamsell.keycloak.authenticator.PhoneNumberConfirmNumber
-com.bayamsell.keycloak.authenticator.PhoneNumberChooseUser
-com.bayamsell.keycloak.authenticator.PhoneNumberSendTan
-com.bayamsell.keycloak.authenticator.PhoneNumberValidateTan
-com.bayamsell.keycloak.authenticator.PhoneNumberUpdateUser
+com.vymalo.keycloak.authenticator.PhoneNumberGetNumber
+com.vymalo.keycloak.authenticator.PhoneNumberConfirmNumber
+com.vymalo.keycloak.authenticator.PhoneNumberChooseUser
+com.vymalo.keycloak.authenticator.PhoneNumberSendTan
+com.vymalo.keycloak.authenticator.PhoneNumberValidateTan
+com.vymalo.keycloak.authenticator.PhoneNumberUpdateUser
